+++ src/main/kotlin/io/csbroker/apiserver/controller/v1/admin/problem/AdminLongProblemController.kt
@@ -7,7 +7,7 @@ import io.csbroker.apiserver.dto.problem.AdminProblemSearchDto
 import io.csbroker.apiserver.dto.problem.longproblem.LongProblemResponseDto
 import io.csbroker.apiserver.dto.problem.longproblem.LongProblemSearchResponseDto
 import io.csbroker.apiserver.dto.problem.longproblem.LongProblemUpsertRequestDto
-import io.csbroker.apiserver.service.problem.AdminLongProblemService
+import io.csbroker.apiserver.service.problem.admin.AdminLongProblemService
 import org.springframework.data.domain.Pageable
 import org.springframework.security.core.userdetails.User
 import org.springframework.web.bind.annotation.GetMapping
