+++ src/test/kotlin/io/csbroker/apiserver/service/problem/admin/AdminLongProblemServiceTest.kt
@@ -1,4 +1,4 @@
-package io.csbroker.apiserver.service.problem
+package io.csbroker.apiserver.service.problem.admin
 
 import io.csbroker.apiserver.auth.ProviderType
 import io.csbroker.apiserver.common.exception.EntityNotFoundException
@@ -10,6 +10,9 @@ import io.csbroker.apiserver.repository.problem.LongProblemRepository
 import io.csbroker.apiserver.repository.problem.ProblemRepository
 import io.csbroker.apiserver.repository.problem.StandardAnswerRepository
 import io.csbroker.apiserver.repository.user.UserRepository
+import io.csbroker.apiserver.service.problem.admin.AdminLongProblemService
+import io.csbroker.apiserver.service.problem.admin.AdminLongProblemServiceImpl
+import io.csbroker.apiserver.service.problem.admin.TagUpserter
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
